backend/requirements.txt
alembic==1.4.2
apiclient==1.0.4
appnope==0.1.0
astroid==2.4.0
attrs==19.3.0
autopep8==1.5.3
awscli==1.18.97
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
boto3==1.13.5
botocore==1.16.5
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==36.0.1
dateparser==0.7.6
decorator==4.4.2
docutils==0.15.2
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-Limiter==1.3.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gcloud==0.18.3
google-api-core==1.16.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-pubsub==1.4.3
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.19.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.3
jmespath==0.9.5
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
limits==1.5.1
lxml==4.6.5
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
nltk==3.6.5
nodeenv==1.3.5
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
pandas==1.3.5
paramiko==2.7.1
parso==0.7.0
pexpect==4.8.0
pg8000==1.16.5
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.11.3
protobuf3-to-dict==0.1.5
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
pylint==2.5.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-env==0.6.2
pytest-mock==3.2.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.4
pyzmq==19.0.1
regex==2020.5.14
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.3
sagemaker==1.65.1.post1
scikit-learn==1.0.2
scipy==1.7.3
scramp==1.2.0
selenium==3.141.0
six==1.11.0
slacker==0.9.65
slacker-log-handler==1.7.1
smdebug-rulesconfig==0.1.4
soupsieve==2.0
SQLAlchemy==1.3.16
testing.common.database==2.0.3
testing.postgresql==1.3.0
texttable==1.6.2
threadpoolctl==2.0.0
tldextract==2.2.2
toml==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.5.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
watchtower==0.8.0
wcwidth==0.2.3
websocket-client==0.57.0
Werkzeug==1.0.1
wordninja==2.0.0
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
